@@ -1,22 +1,23 @@
-import {Injectable} from '@angular/core'
-import * as Fuse from 'fuse.js'
-import FuseOptions = Fuse.FuseOptions;
+import {Injectable} from '@angular/core';
+import * as Fuse from 'fuse.js';
+import {FuseOptions} from 'fuse.js'
 
-import _set = require('lodash.set');
-import _get = require('lodash.get');
+import {set as _set} from 'lodash-es';
+import {get as _get} from 'lodash-es';
 
 
-export interface AngularFusejsOptions extends FuseOptions {
+export interface AngularFusejsOptions<T> extends FuseOptions<T> {
   supportHighlight?: boolean;
   fusejsHighlightKey?: string;
   fusejsScoreKey?: string;
-  minSearchTermLength?: number;
+  minSearchTermLength?: number; // = 0;
   maximumScore?: number;
+  highlightTag?: string;
 }
 
 @Injectable()
-export class FusejsService {
-  private defaultOptions: AngularFusejsOptions = {
+export class FusejsService<T> {
+  private defaultOptions: AngularFusejsOptions<T> = {
     supportHighlight: true,
     shouldSort: false,
     threshold: 0.6,
@@ -30,11 +31,13 @@ export class FusejsService {
     fusejsScoreKey: 'fuseJsScore',
   };
 
-  searchList(list: Array<any>, searchTerms: string, options: AngularFusejsOptions = {}) {
-    const fuseOptions: AngularFusejsOptions = Object.assign({}, this.defaultOptions, options);
-    let result = [];
+  searchList(list: Array<T>, searchTerms: string, options: AngularFusejsOptions<T> = {}) {
+    // https://stackoverflow.com/questions/35959372/property-assign-does-not-exist-on-type-objectconstructor
+    // TODO : remove (<any>Object) hack by using right lib or polyfill ?
+    const fuseOptions: AngularFusejsOptions<T> = (<any>Object).assign({}, this.defaultOptions, options);
+    let result:any = [];
 
-    if (searchTerms && searchTerms.length >= fuseOptions.minSearchTermLength) {
+    if (searchTerms && searchTerms.length >= (fuseOptions?.minSearchTermLength || 0)) {
       if (fuseOptions.supportHighlight) {
         fuseOptions.includeMatches = true;
       }
@@ -49,7 +52,7 @@ export class FusejsService {
 
       if (fuseOptions.supportHighlight) {
         result.forEach((element) => {
-          element[fuseOptions.fusejsHighlightKey] = this.deepClone(element);
+          element[fuseOptions.fusejsHighlightKey || '_'] = this.deepClone(element);
         });
       }
     }
@@ -67,36 +70,38 @@ export class FusejsService {
     return _out;
   }
 
-  private handleHighlight(result, options: AngularFusejsOptions) {
+  private handleHighlight(result, options: AngularFusejsOptions<T>) {
     if (options.maximumScore && options.includeScore) {
       result = result.filter((matchObject) => {
-        return matchObject.score <= options.maximumScore;
+        return matchObject.score <= (options.maximumScore||0);
       })
     }
 
     return result.map((matchObject) => {
       const item = this.deepClone(matchObject.item);
-      item[options.fusejsHighlightKey] = this.deepClone(item);
-      item[options.fusejsScoreKey] = matchObject.score;
+      item[options.fusejsHighlightKey || "_"] = this.deepClone(item);
+      item[options.fusejsScoreKey || "_"] = matchObject.score;
       for (let match of matchObject.matches) {
         const indices: number[][] = match.indices;
 
         let highlightOffset: number = 0;
 
         let key: string = match.key;
-        if(_get(item[options.fusejsHighlightKey], key).constructor === Array) {
+        if(_get(item[options.fusejsHighlightKey || "_"], key).constructor === Array) {
           key += `[${match.arrayIndex}]`
         }
 
         for (let indice of indices) {
-          let initialValue: string = _get(item[options.fusejsHighlightKey], key) as string;
+          let initialValue: string = _get(item[options.fusejsHighlightKey || "_"], key) as string;
 
           const startOffset = indice[0] + highlightOffset;
           const endOffset = indice[1] + highlightOffset + 1;
+          const tagStart = "<" + (options.highlightTag ?? "em") + ">";
+          const tagEnd = "</" + (options.highlightTag ?? "em") + ">";
           let highlightedTerm = initialValue.substring(startOffset, endOffset);
-          let newValue = initialValue.substring(0, startOffset) + '<em>' + highlightedTerm + '</em>' + initialValue.substring(endOffset);
-          highlightOffset += '<em></em>'.length;
-          _set(item[options.fusejsHighlightKey], key, newValue);
+          let newValue = initialValue.substring(0, startOffset) + tagStart + highlightedTerm + tagEnd + initialValue.substring(endOffset);
+          highlightOffset += (tagStart + tagEnd).length;
+          _set(item[options.fusejsHighlightKey || "_"], key, newValue);
         }
       }
 
